pyproject.toml
[project]
name = "entraid-mcp-server"
version = "0.2.0"
description = "FastMCP server for Microsoft Graph / Entra ID, using Azure DefaultAzureCredential by default."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "baoduy" },
]
keywords = [
    "mcp",
    "fastmcp",
    "microsoft-graph",
    "entra-id",
    "azure",
    "azure-identity",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration :: Authentication/Directory",
]
dependencies = [
    "azure-core>=1.39.0",
    "azure-identity>=1.25.3",
    "fastmcp>=2.10.0",
    "msgraph-core>=1.3.8",
    "msgraph-sdk>=1.56.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
entraid-mcp-server = "msgraph_mcp_server.server:main"

[project.urls]
Homepage = "https://github.com/baoduy/entraid-mcp-server"
Repository = "https://github.com/baoduy/entraid-mcp-server"
Issues = "https://github.com/baoduy/entraid-mcp-server/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/msgraph_mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
    "src/msgraph_mcp_server",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]
